+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Improved import merge: consider last modified date, merge labels, do not merge if reminders differ (#49).
 - More note text is now shown in reminder notification.
 - Select newly created label when created from main screen.
+- Focus text note content if background is clicked.
 - Fixed export not completely overwriting JSON file if new content smaller.
 - Fixed crash during search if maximum lines shown in preview is 1 or 2 (#47).
 - Fixed crash during search if only negative term is used (e.g. `-a`) (#47).